backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
pdfplumber==0.11.4
anthropic==0.40.0
python-multipart==0.0.12
pydantic==2.9.2
python-dotenv==1.0.1
